Make the bash script resilient to having no log file

+++ orchestration/scripts/move-logs-to-bucket.sh
@@ -1,9 +1,14 @@
 set -euo pipefail
 
 declare -r LOGS_DIR=../../transformation/logs
+declare -r LOG_FILE=${LOGS_DIR}/errors.log
 declare -r NEW_FILENAME=${LOGS_DIR}/${TIMESTAMP}.log
 
-# rename the file to the start time of the import, move to a bucket
-mv ${LOGS_DIR}/errors.log ${NEW_FILENAME}
-gsutil cp ${NEW_FILENAME} ${BUCKET_NAME}/errors/
-rm ${NEW_FILENAME}
+if test -f "$LOG_FILE"; then
+  # rename the file to the start time of the import
+  mv ${LOG_FILE} ${NEW_FILENAME}
+  # move the file to a bucket
+  gsutil cp ${NEW_FILENAME} ${BUCKET_NAME}/errors/
+  rm ${NEW_FILENAME}
+fi
+
